+++ docs/discord-bot-commands.md
