--- example/collaboration.md
@@ -1,1 +1,0 @@
-collaboration
