+++ knowledge-king/src/main/kotlin/tw/waterballsa/utopia/knowledgeking/domain/KinoftheQuizBase.kt
