+++ main/src/main/resources/wsa.beta.properties
@@ -1,0 +1,2 @@
+guild-id=1038654765896310804
+unlock-entry-message-id=1038667013259792455
