+++ main/src/main/resources/wsa.prod.properties
@@ -1,0 +1,2 @@
+guild.id=937992003415838761
+unlock-entry-message-id=1042776102734663740
