+++ two-sum/pom.xml
@@ -1,0 +1,27 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+  <parent>
+    <artifactId>root</artifactId>
+    <groupId>tw.waterballsa.utopia</groupId>
+    <version>${revision}</version>
+  </parent>
+
+  <artifactId>two-sum</artifactId>
+
+  <properties>
+
+  </properties>
+
+  <dependencies>
+
+    <dependency>
+      <groupId>tw.waterballsa.utopia</groupId>
+      <artifactId>commons</artifactId>
+    </dependency>
+  </dependencies>
+
+
+</project>
